Refactor and Improved code Quality- Fix typos, removed unused imports, code more cleaner, update more readable readme.md and update junit testing.
