+++ futures_api_example_java/readme.md
@@ -1,25 +1,15 @@
 ## Getting started
 
-#### Public API
+#### How to run this project
 
-* junit
-run MarketTest
+* To create a new self executable jar 
+  * run the command on the terminal > mvn clean install
+  * run self executable jar > java -jar <jarname> e.g. bitrue-future-client-0.9.0-SNAPSHOT-jar-with-dependencies.jar
+  * run unit test > mvn test
+  * skip unit test > mvn clean install -DskipTests
   
 
-#### Trade & Account API
-
-``` java
-
-public class PrivateConfig {
-
-    public static final String API_KEY = "XXXXXX";
-    public static final String SECRET_KEY = "XXXXXX";
-
-}
-
-# junit:
-TradeTest
-AccountTest
-```
+#### API key and signature
+* See the class PrivateConfig.java, add your Bittrue APIKey and Signature
