+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/PrivateConfig.java
@@ -1,7 +1,6 @@
 package com.bitrue.futures.sdk.client;
 
 public class PrivateConfig {
-
     public static final String API_KEY = "yyy";
     public static final String SECRET_KEY = "xxx";
 }
