+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/RequestOptions.java
@@ -1,5 +1,6 @@
 package com.bitrue.futures.sdk.client;
 
+import com.bitrue.futures.sdk.client.constant.FuturesApiConstants;
 import com.bitrue.futures.sdk.client.exception.BitrueApiException;
 
 import java.net.URL;
