+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/constant/FuturesApiConstants.java
@@ -1,4 +1,4 @@
-package com.bitrue.futures.sdk.client;
+package com.bitrue.futures.sdk.client.constant;
 
 import org.apache.commons.lang.builder.ToStringStyle;
