+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/exception/BitrueApiException.java
@@ -2,15 +2,10 @@ package com.bitrue.futures.sdk.client.exception;
 
 public class BitrueApiException extends RuntimeException {
 
-    /**
-     *
-     */
     private static final long serialVersionUID = 4360108982268949009L;
     public static final String RUNTIME_ERROR = "RuntimeError";
     public static final String INPUT_ERROR = "InputError";
     public static final String KEY_MISSING = "KeyMissing";
-    public static final String SYS_ERROR = "SysError";
-    public static final String SUBSCRIPTION_ERROR = "SubscriptionError";
     public static final String ENV_ERROR = "EnvironmentError";
     public static final String EXEC_ERROR = "ExecuteError";
     private final String errCode;
