+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/impl/RestApiRequest.java
@@ -3,8 +3,6 @@ package com.bitrue.futures.sdk.client.impl;
 import okhttp3.Request;
 
 public class RestApiRequest <T>{
-
     public Request request;
-
     RestApiJsonParser<T> jsonParser;
 }
