+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/account/Account.java
@@ -6,6 +6,9 @@ import lombok.Data;
 import java.math.BigDecimal;
 import java.util.List;
 
+/**
+ * Account details
+ */
 @Data
 @Builder
 public class Account {
