+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/account/PositionVO.java
@@ -5,6 +5,9 @@ import lombok.Data;
 
 import java.util.List;
 
+/**
+ * Position Value Object details
+ */
 @Data
 @Builder
 public class PositionVO {
