+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/enums/Interval.java
@@ -11,7 +11,7 @@ public enum Interval {
     /**
      * value
      */
-    private String value;
+    private final String value;
 
     Interval(String v){
         this.value = v;
