+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/enums/OrderType.java
@@ -1,6 +1,5 @@
 package com.bitrue.futures.sdk.client.model.enums;
 
 public enum OrderType {
-
-    LIMIT, MARKET;
+    LIMIT, MARKET
 }
