+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/enums/PositionActiion.java
@@ -1,5 +1,5 @@
 package com.bitrue.futures.sdk.client.model.enums;
 
 public enum PositionActiion {
-    OPEN, CLOSE;
+    OPEN, CLOSE
 }
