+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/enums/TimeInForce.java
@@ -1,6 +1,5 @@
 package com.bitrue.futures.sdk.client.model.enums;
 
 public enum TimeInForce {
-
     IOC, FOK, POST_ONLY, LIMIT, MARKET;
 }
