+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/market/ContractInfo.java
@@ -1,7 +1,6 @@
 package com.bitrue.futures.sdk.client.model.market;
 
 import com.bitrue.futures.sdk.client.model.enums.ContractSide;
-import com.oracle.webservices.internal.api.databinding.DatabindingMode;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -9,7 +8,7 @@ import lombok.EqualsAndHashCode;
 import java.math.BigDecimal;
 
 /**
- *
+ * Contract details
  * @author superatom
  */
 @Data
