+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/market/OrderBookEntry.java
@@ -1,6 +1,6 @@
 package com.bitrue.futures.sdk.client.model.market;
 
-import com.bitrue.futures.sdk.client.FuturesApiConstants;
+import com.bitrue.futures.sdk.client.constant.FuturesApiConstants;
 import lombok.Builder;
 import lombok.Data;
 import org.apache.commons.lang.builder.ToStringBuilder;
