+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/market/PriceChangeTicker.java
@@ -5,6 +5,9 @@ import lombok.Data;
 
 import java.math.BigDecimal;
 
+/**
+ * Price Change Ticker
+ */
 @Builder
 @Data
 public class PriceChangeTicker {
