+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/market/ServerTime.java
@@ -3,6 +3,9 @@ package com.bitrue.futures.sdk.client.model.market;
 import lombok.Builder;
 import lombok.Data;
 
+/**
+ * Server time details
+ */
 @Data
 @Builder
 public class ServerTime {
