+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/model/trade/Order.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import java.math.BigDecimal;
 
 /**
+ * Order Details
  * @author superatom
  */
 @Data
