+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/utils/Handler.java
@@ -2,6 +2,5 @@ package com.bitrue.futures.sdk.client.utils;
 
 @FunctionalInterface
 public interface Handler<T> {
-
     void handle(T t);
 }
