+++ futures_api_example_java/src/main/java/com/bitrue/futures/sdk/client/utils/JsonWrapper.java
@@ -220,7 +220,7 @@ public class JsonWrapper {
 
     public JsonWrapperArray getJsonArray(String name) {
         checkMandatoryField(name);
-        JSONArray array = null;
+        JSONArray array;
         try {
             array = json.getJSONArray(name);
         } catch (Exception e) {
